test-project/requirements.txt
# Core dependencies for Flash
tetra-rp>=0.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
